pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "od-draw"
authors = [{name = "AJ Carter", email = "aj@example.com"}]
description = "A Python library for creating diagrams with multiple backend support (SVG, PNG, Draw.io)"
keywords = ["diagrams", "draw.io", "svg", "visualization"]
readme = "README.md"
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "pillow >= 10.0.0",
    "click >= 8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest >= 7.0.0",
    "mypy >= 1.0.0",
    "ruff >= 0.1.0",
]
docs = [
    "sphinx >= 7.0.0",
    "furo >= 2023.0.0",
]

[project.urls]
Homepage = "https://github.com/ajcarter/od-draw"
Issues = "https://github.com/ajcarter/od-draw/issues"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
